requirements.txt
beautifulsoup4==4.12.2
certifi==2022.12.7
charset-normalizer==3.1.0
DateTime==5.1
idna==3.4
pytz==2023.3
regex==2023.3.23
requests==2.28.2
soupsieve==2.4.1
urllib3==1.26.15
zope.interface==6.0
tqdm
python-telegram-bot==13.8.1
